# starmap-contracts/openapi.yaml
openapi: "3.0.3"
info:
  title: "星图 StarMap API"
  description: "人才能力星云导航系统 - 后端 API 契约"
  version: "1.0.0"
  contact:
    name: "StarMap Team"
  x-summary: "人才技能图谱知识系统全量 API"
servers:
  - url: "http://localhost:8000/api/v1"
    description: "本地开发环境"
  - url: "https://api.starmap.dev/api/v1"
    description: "评审环境"
security:
  - BearerAuth: []
paths:
...
  /datasources/{source_id}:
    get:
      operationId: getDatasourceDetail
      summary: "数据源详情"
      description: "获取指定数据源的详细配置和统计"
      tags: [数据源管理]
      parameters:
        - name: source_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: "数据源详情"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataSourceResponse"
        "404":
          description: "数据源不存在"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      operationId: updateDatasource
      summary: "更新数据源"
      description: "更新数据源的权威度评分等配置"
      tags: [数据源管理]
      parameters:
        - name: source_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DataSourceUpdateRequest"
      responses:
        "200":
          description: "更新成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataSourceResponse"
        "404":
          description: "数据源不存在"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /pipeline/runs:
    get:
      operationId: listPipelineRuns
      summary: "流水线运行记录"
      description: "分页查询流水线运行历史"
      tags: [数据流水线]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: status
          in: query
          schema:
            type: string
          description: "按状态筛选"
      responses:
        "200":
          description: "运行记录列表"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PipelineRunResponse"
  /pipeline/runs/{run_id}:
    get:
      operationId: getPipelineRunDetail
      summary: "流水线运行详情"
      description: "获取单次流水线运行的完整详情，包含各阶段状态"
      tags: [数据流水线]
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: "运行详情"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineRunResponse"
        "404":
          description: "运行记录不存在"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /pipeline/runs/{run_id}/cancel:
    post:
      operationId: cancelPipelineRun
      summary: "取消流水线运行"
      description: "软取消 + Redis STOP flag + Celery 阶段开始时检查"
      tags: [数据流水线]
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: "取消成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CancelResponse"
  /pipeline/runs/{run_id}/retry:
    post:
      operationId: retryPipelineStage
      summary: "重试失败阶段"
      description: "重试指定流水线运行的失败阶段（断点续跑）"
      tags: [数据流水线]
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [stage_name]
              properties:
                stage_name:
                  type: string
      responses:
        "200":
          description: "重试成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineRunResponse"
        "404":
          description: "运行记录不存在"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /pipeline/runs/{run_id}/resume:
    post:
      operationId: resumePipelineRun
      summary: "断点续跑"
      description: "重置所有失败阶段并继续执行"
      tags: [数据流水线]
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: "续跑成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineRunResponse"
        "404":
          description: "运行记录不存在"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /pipeline/trigger:
    post:
      operationId: triggerPipeline
      summary: "触发流水线"
      description: "手动触发一次数据流水线运行（DAG 执行，支持阶段选择）"
      tags: [数据流水线]
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TriggerRequest"
      responses:
        "200":
          description: "流水线已触发"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TriggerResponse"
...
  /pipeline/events-poll:
    get:
      operationId: pollPipelineEvents
      summary: "流水线事件轮询"
      description: "SSE 轮询回退，返回最近事件数组"
      tags: [数据流水线]
      parameters:
        - name: since
          in: query
          schema:
            type: number
            default: 0.0
          description: "Unix 时间戳筛选"
      responses:
        "200":
          description: "最近事件"
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
  /pipeline/schedules:
    get:
      operationId: listSchedules
      summary: "定时调度列表"
      description: "列出所有定时调度"
      tags: [数据流水线]
      responses:
        "200":
          description: "调度列表"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScheduleResponse"
    post:
      operationId: createSchedule
      summary: "创建定时调度"
      description: "创建定时调度（创建时计算 next_run_at）"
      tags: [数据流水线]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleCreateRequest"
      responses:
        "200":
          description: "调度已创建"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleResponse"
  /pipeline/schedules/{schedule_id}:
    put:
      operationId: updateSchedule
      summary: "更新定时调度"
      description: "更新定时调度配置"
      tags: [数据流水线]
      parameters:
        - name: schedule_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleCreateRequest"
      responses:
        "200":
          description: "调度已更新"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleResponse"
        "404":
          description: "调度不存在"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteSchedule
      summary: "删除定时调度"
      description: "删除指定定时调度"
      tags: [数据流水线]
      parameters:
        - name: schedule_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: "调度已删除"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "deleted"
        "404":
          description: "调度不存在"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /pipeline/schedules/{schedule_id}/trigger:
    post:
      operationId: triggerSchedule
      summary: "手动触发调度"
      description: "读取调度配置并触发流水线"
      tags: [数据流水线]
      parameters:
        - name: schedule_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: "调度已触发"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TriggerResponse"
        "404":
          description: "调度不存在"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /pipeline/config:
    get:
      operationId: getPipelineConfig
      summary: "获取流水线配置"
      description: "获取流水线配置（超时/并发/重试）"
      tags: [数据流水线]
      responses:
        "200":
          description: "流水线配置"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineConfigResponse"
    put:
      operationId: updatePipelineConfig
      summary: "更新流水线配置"
      description: "更新流水线配置（写入 .env 或运行时覆盖）"
      tags: [数据流水线]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PipelineConfigUpdateRequest"
      responses:
        "200":
          description: "配置已更新"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineConfigResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT Bearer token obtained from authentication endpoint"
  schemas:
...
    Error:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          description: "错误描述"
          example: "岗位不存在"
        code:
          type: string
          description: "错误码（可选，部分端点返回）"
          example: "POSITION_NOT_FOUND"
        timestamp:
          type: string
          format: date-time
          description: "错误发生时间（可选）"

    # ──────────────────────── 数据大屏 ────────────────────────
...
    DataSourceResponse:
      type: object
      required: [id, name, source_type, authority_score, status]
      properties:
        id:
          type: string
        name:
          type: string
        source_type:
          type: string
        authority_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
        status:
          type: string
        last_crawl_at:
          type: string
          format: date-time
          nullable: true
        total_records:
          type: integer
        valid_records:
          type: integer
        duplicate_rate:
          type: number
          format: float
        avg_quality_score:
          type: number
          format: float
        config:
          type: object
          additionalProperties: true
    DataSourceUpdateRequest:
      type: object
      properties:
        authority_score:
          type: number
          nullable: true
          minimum: 0
          maximum: 1
        status:
          type: string
          nullable: true
        config:
          type: object
          nullable: true
...
    TriggerRequest:
      type: object
      properties:
        run_type:
          type: string
          default: "full"
        selected_stages:
          type: array
          items:
            type: string
          nullable: true
...